+++ Qbert/SingleplayerLevelScene.cpp
@@ -37,7 +37,7 @@ dae::SingleplayerLevelScene::SingleplayerLevelScene(int level, int score)
 
 void dae::SingleplayerLevelScene::OnSetup() {
 	const auto& levels{ Qbert::GetInstance().GetLevelInfo() };
-	if (m_level >= levels.size()) {
+	if (m_level >= static_cast<int>(levels.size())) {
 		std::cout << "Level does not exist, idiot" << std::endl;
 		return;
 	}
